Altered Login Logo Function to Allow a Height Function to be Used
The height was hard coded into the logo login function and therefore
the height is now in a seperate function. This means that a login logo
can be added to the child themes images folder and then if the height
of the logo needs changing only the height function needs to be
declared in the child theme functions file.